chore: address desktop ui review feedback
Agent-Logs-Url: https://github.com/hussainshahzad350/phone_shop_pos/sessions/a3f1acee-b6d9-4fed-8157-e6378a4cf0f9

+++ lib/core/widgets/desktop_components.dart
@@ -439,9 +439,7 @@ class _DesktopContentArea extends StatelessWidget {
         return Align(
           alignment: Alignment.topCenter,
           child: SizedBox(
-            width: constraints.maxWidth
-                .clamp(0.0, _kDesktopContentMaxWidth)
-                .toDouble(),
+            width: constraints.maxWidth.clamp(0.0, _kDesktopContentMaxWidth),
             child: child,
           ),
         );
